# expedite eye and zeros method