`eye` and `zeros` methods have been expedited by several factors.

Fixes #21003


#### Brief description of what is fixed or changed
Improved `zeros` by using lists instead, and for `eye`, we are using `zeros` and then assigning `1` to every diagonal element.

Credits: @oscarbenjamin for suggesting the idea and code.

#### Release Notes
<!-- BEGIN RELEASE NOTES -->
* matrices
    * The eye and zero functions have been made much faster for large matrices.
<!-- END RELEASE NOTES -->