@hyadav2k 2021-02-27
@oscarbenjamin, this diff works and it speeds up `eye` almost by a factor of double.
This also passes the test cases.
```
diff --git a/sympy/matrices/common.py b/sympy/matrices/common.py
index 28a9269588..74f609e192 100644
--- a/sympy/matrices/common.py
+++ b/sympy/matrices/common.py
@@ -772,11 +772,10 @@ def entry(i, j):

     @classmethod
     def _eval_eye(cls, rows, cols):
-        from sympy.matrices import zeros
-        M = zeros(rows, cols)
-        for i in range(rows):
-            M[i*(1+rows)] = cls.one
-        return cls._new(M)
+        from sympy.matrices.dense import Matrix
+        vals = [cls.zero]*(rows*cols)
+        vals[::cols+1] = [cls.one]*min(rows, cols)
+        return cls._new(Matrix._new(rows, cols, vals, copy=False))

     @classmethod
     def _eval_jordan_block(cls, rows, cols, eigenvalue, band='upper'):
diff --git a/sympy/matrices/dense.py b/sympy/matrices/dense.py
index c2484eaefc..5c405bc495 100644
--- a/sympy/matrices/dense.py
+++ b/sympy/matrices/dense.py
@@ -305,11 +305,7 @@ def _new(cls, *args, copy=True, **kwargs):
             # It should be used directly without creating a copy.
             if len(args) != 3:
                 raise TypeError("'copy=False' requires a matrix be initialized as rows,cols,[list]")
-            else:
-                rows, cols, flat_list = args
-                if rows < 0 or cols < 0:
-                    raise ValueError("Cannot create a {} x {} matrix. "
-                                     "Both dimensions must be positive".format(rows, cols))
+            rows, cols, flat_list = args
         else:
             rows, cols, flat_list = cls._handle_creation_inputs(*args, **kwargs)
             flat_list = list(flat_list) # create a shallow copy
```
Some exapmles:
```
In [1]: eye(-4,-3)
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
<ipython-input-1-bc5ca6a87586> in <module>
----> 1 eye(-4,-3)

~\sympy\sympy\matrices\dense.py in eye(*args, **kwargs)
    916     """
    917
--> 918     return Matrix.eye(*args, **kwargs)
    919
    920

~\sympy\sympy\matrices\common.py in eye(kls, rows, cols, **kwargs)
    980         rows, cols = as_int(rows), as_int(cols)
    981
--> 982         return klass._eval_eye(rows, cols)
    983
    984     @classmethod

~\sympy\sympy\matrices\common.py in _eval_eye(cls, rows, cols)
    775         from sympy.matrices.dense import Matrix
    776         vals = [cls.zero]*(rows*cols)
--> 777         vals[::cols+1] = [cls.one]*min(rows, cols)
    778         return cls._new(Matrix._new(rows, cols, vals, copy=False))
    779

ValueError: attempt to assign sequence of size 0 to extended slice of size 6

In [2]: eye(-4,3)
Out[2]: []

In [3]: eye(4,-3)
Out[3]: []

In [4]: zeros(-4,4)
Out[4]: []

In [5]: zeros(-4,-3)
Out[5]: []

In [6]: zeros(4,-3)
Out[6]: []

```