@oscarbenjamin 2021-03-12
Can you squash the commits into one?

How do the timings compare with master now?