@hyadav2k 2021-03-13
This implementation does not raise any error for examples like:
```
In [1]: zeros(3, -1)
Out[1]: []

In [2]: zeros(-3, -1)
Out[2]: []

In [3]: zeros(-3, 1)
Out[3]: []

In [4]: eye(3, -1)
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
<ipython-input-4-06bed9c6d91c> in <module>
----> 1 eye(3, -1)

~\sympy\sympy\matrices\dense.py in eye(*args, **kwargs)
    916     """
    917
--> 918     return Matrix.eye(*args, **kwargs)
    919
    920

~\sympy\sympy\matrices\common.py in eye(kls, rows, cols, **kwargs)
    980         rows, cols = as_int(rows), as_int(cols)
    981
--> 982         return klass._eval_eye(rows, cols)
    983
    984     @classmethod

~\sympy\sympy\matrices\common.py in _eval_eye(cls, rows, cols)
    775         from sympy.matrices.dense import Matrix
    776         vals = [cls.zero]*(rows*cols)
--> 777         vals[::cols+1] = [cls.one]*min(rows, cols)
    778         return cls._new(Matrix._new(rows, cols, vals, copy=False))
    779

ValueError: slice step cannot be zero

In [5]: eye(-3, -1)
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
<ipython-input-5-a10d9d9837a3> in <module>
----> 1 eye(-3, -1)

~\sympy\sympy\matrices\dense.py in eye(*args, **kwargs)
    916     """
    917
--> 918     return Matrix.eye(*args, **kwargs)
    919
    920

~\sympy\sympy\matrices\common.py in eye(kls, rows, cols, **kwargs)
    980         rows, cols = as_int(rows), as_int(cols)
    981
--> 982         return klass._eval_eye(rows, cols)
    983
    984     @classmethod

~\sympy\sympy\matrices\common.py in _eval_eye(cls, rows, cols)
    775         from sympy.matrices.dense import Matrix
    776         vals = [cls.zero]*(rows*cols)
--> 777         vals[::cols+1] = [cls.one]*min(rows, cols)
    778         return cls._new(Matrix._new(rows, cols, vals, copy=False))
    779

ValueError: slice step cannot be zero

In [6]: eye(-3, 1)
Out[6]: []
```
This is why it fails the test.
One of the way to tackle is adding the code from the previous commit as:
```
else:
        rows, cols, flat_list = args
        if rows < 0 or cols < 0:
            raise ValueError("Cannot create a {} x {} matrix. "
                             "Both dimensions must be positive".format(rows, cols))
```
>
>
For the timings:
```
On branch:
In [1]: %time ok = zeros(1000)
Wall time: 17 ms

In [2]: %time ok = eye(1000)
Wall time: 19.9 ms
```
```
On master:
In [1]: %time ok = zeros(1000)
Wall time: 2.02 s

In [2]: %time ok = eye(1000)
Wall time: 2.31 s
```