@oscarbenjamin 2021-03-13
It's better to add those checks in the `eye` and `zeros` functions.