@hyadav2k 2021-03-13
The timings seem to remain almost the same after adding the commit.

@oscarbenjamin, if this looks good, I will squash the commits into one.
Thanks.